Prevent a PHP 8.0 deprecated notice about optional function argument
The `bp_admin_get_wp_nav_menu_items()` funciton was introduced in 7.0.0 to fix an issue with the BuddyPress Members WP Nav Menu items. It was wrongly using an optional argument before a required one. As PHP 8.0 no longer accepts this practice, both arguments of the function will be required 😇.

Fixes #8390


git-svn-id: https://buddypress.svn.wordpress.org/trunk@12780 cdf35c40-ae34-48e0-9cc9-0c9da1808c22