xProfile: make sure a Field Type can enforce field visibility
When saving a field having a Field Type where the custom visibility feature is disabled and the visibility property is set, use this visibility to enforce the field visibility. This part was forgotten in #7162 and it is needed by the Checkbox Acceptance field to make sure only the logged in user can see they agreed to the page requiring their acceptance.

Props mahdiar, vapvarun

See #4747



git-svn-id: https://buddypress.svn.wordpress.org/trunk@12881 cdf35c40-ae34-48e0-9cc9-0c9da1808c22