Make sure saving empty WP xProfile fields does generate an error
WP xProfile field types are allowing their values to be empty. In such a case the `xprofile_set_field_data()` is using an array as the value to be saved. This is done to avoid the value of certain types of fields (checkboxes, multiselects) to be overwritten by the default value on next edit.

For WP xProfile field types, we simply need to return an empty string in this case.

Fixes #8494 (trunk)



git-svn-id: https://buddypress.svn.wordpress.org/trunk@12978 cdf35c40-ae34-48e0-9cc9-0c9da1808c22