Avoid a PHP8 warning when xProfile user visibility levels are empty
When the record does not exist yet into the usermeta table, getting the `'bp_xprofile_visibility_levels'` usermeta returns an empty string while we're expecting an array. As we use the expected array to reset the visibility level if the administrator enforced a specific visibility to the field, it's causing the PHP8 warning. To fix it, we are now casting the `'bp_xprofile_visibility_levels'` usermeta as soon as we get it.

Props johnjamesjacoby

See #8503 (branch 8.0)



git-svn-id: https://buddypress.svn.wordpress.org/branches/8.0@13011 cdf35c40-ae34-48e0-9cc9-0c9da1808c22