Activity: update the nonce used by the Activity Reply JS Fallback
If JavaScript is unavailable (whether it was disabled or it errored), posting an activity uses the `bp_activity_action_post_comment()` PHP function after a page load. Unlike the Ajax action, the nonce used into this function has not been updated in [12735]. The main goal of this commit is to fix this oversight. The secondary goal is to improve code formatting.

Fixes #8545 (branch 9.0) 



git-svn-id: https://buddypress.svn.wordpress.org/branches/9.0@13042 cdf35c40-ae34-48e0-9cc9-0c9da1808c22