Avoid 2 superfluous private message notification queries
When viewing the single view of the thread, the potential notifications about it are marked as read into the `messages_action_conversation()` function. We don't need to do it again when looping into the thread messages.

This was revealed by a database error which was appearing when a user who had received more than 1 message from two different threads was directly viewing the single view of one of the threads.

Props mike80222, oztaser

Fixes #8676 (10.0 branch)



git-svn-id: https://buddypress.svn.wordpress.org/branches/10.0@13270 cdf35c40-ae34-48e0-9cc9-0c9da1808c22