Avoid PHP 8.1 ArrayAccess Return Type deprecation notices
The `BP_Core_BP_Nav_BackCompat` is implementing the ArrayAccess class to provide backward compatibility to BuddyPress plugins directly modifying the `buddypress()->bp_nav` global since BuddyPress 2.3.

PHP 8.1 added a deprecation notice if a class declaration has an incompatible return type (or if the return type lacks: which is the case for `BP_Core_BP_Nav_BackCompat`) with the parent it extends.

As BuddyPress supports PHP back to version 5.6, we need to use the `#ReturnTypeWillChange` attribute to maintain compatibility with this version and avoid the deprecation notice in PHP 8.x.

Fixes #8726



git-svn-id: https://buddypress.svn.wordpress.org/trunk@13311 cdf35c40-ae34-48e0-9cc9-0c9da1808c22