Edit translator comments of problematic strings
In 11.0 we have removed 1 placeholder to 2 strings used to display the 
activity action for published posts. Issues were raised about the fact 
some polyglots teams did not update corresponding translations leaving the 
removed placeholder in their translated strings. As these two strings are 
used inside `sprintf()` function it resulted in fatal errors for the 
corresponding locales.

We hope simply improving the comment will avoid such mistakes raising 
translators attention. The result of this commit should avoid the string 
to be translated as long as the polyglots team haven't updated their 
translations.

It is a temporary fix, the right one should be to create a "sanitization" 
function wrapping `sprintf()` to make sure expected placeholders are found 
into the translated strings.

Although we think this kind of function should be made available by 
WordPress, in [https://core.trac.wordpress.org/ticket/55776#comment:3 this 
ticket] the issue was diagnosed as invalid. It's probably ok when you have 
a lot of polyglots contributors like WordPress has, **it's a different 
situation for plugins**: we will work on preventing these potential fatals 
in the next major release (12.0).

Props 1benk1, webmasterfreya

Fixes #8821 (branch 11.0)



git-svn-id: https://buddypress.svn.wordpress.org/branches/11.0@13420 cdf35c40-ae34-48e0-9cc9-0c9da1808c22