Remove `bp-default` altogether from the BuddyPress package
This theme is no more maintained (except for security issues) and will be now available from the BP Classic plugin. Almost 10 years after this announcement: https://bpdevel.wordpress.com/2013/11/13/the-future-of-the-bp-default-theme/ we are finally "removing `bp-default` altogether from the BuddyPress package".

Thanks for the great job people who contributed to it accomplished. It helped BuddyPress a lot to provide a flexible and powerful foundation for building BuddyPress sites.

Fixes #8904
Closes https://github.com/buddypress/buddypress/pull/105



git-svn-id: https://buddypress.svn.wordpress.org/trunk@13490 cdf35c40-ae34-48e0-9cc9-0c9da1808c22