Make sure Edit Photo / Cover image slugs can be customized
Wait for the navigation to be fully set before eventually moving the Member's profile sub nav items into the Member's xProfile sub nav. This make sure the slugs to edit profile photo and cover image can be customized when the BP root site is not the same than the main site of a network on multisite configurations.

Fixes #9050 (branch 12.0)



git-svn-id: https://buddypress.svn.wordpress.org/branches/12.0@13683 cdf35c40-ae34-48e0-9cc9-0c9da1808c22