Add directory post type entries to "page on front" choices if needed
When the `rewrites` URL parser is in use, to preserve the possibility to use  a BuddyPress directory as the site's front page, we need to append the corresponding post type IDs to the available pages listed into the WP "page on front" reading setting as BP Directories are stored as `buddypress` post types instead of the `page` one.

On the other hand, when the `legacy` URL parser is in use (eg: BP Classic is active), to prevent duplicates, it's important to leave the "page on front" reading setting choices the way WP sets it as BP Directories are stored as `page` post types.

Props emaralive

Fixes #9052 (branch 12.0)



git-svn-id: https://buddypress.svn.wordpress.org/branches/12.0@13688 cdf35c40-ae34-48e0-9cc9-0c9da1808c22