Deprecate `bp_use_wp_admin_bar()` & `BP_USE_WP_ADMIN_BAR`
This function and constant were useful when BuddyPress was including an alternative toolbar to the WordPress Admin Bar called the "BuddyBar". The BuddyBar was removed in [12893] but we forgot to deprecate `bp_use_wp_admin_bar()` & `BP_USE_WP_ADMIN_BAR`.  This commit is finally achieving the step we missed almost 3 years ago.

Props espellcaste, emaralive

Fixes #9104
Closes https://github.com/buddypress/buddypress/pull/234



git-svn-id: https://buddypress.svn.wordpress.org/trunk@13741 cdf35c40-ae34-48e0-9cc9-0c9da1808c22