Remove the tool to repair the last active members
This tool was designed to be used in case a `last_activity` user meta migration went wrong during 2.0.0 uprade process. Since this version we stopped using this user meta in favor of a logging "hidden" activity entry.

As `bp_admin_repair_last_activity()` now only remove all trace of last active members without regenarating it. We are deprecating and stop using this function.