Remove the tool to repair the members last activity tracker
In version 2.0 we decided to stop using the `last_activity` user meta in favor of tracing this information as an activity type having the same name. Meaning we stopped inserting this user meta for new members or any member of a community if the first BuddyPress installed version is upper than 1.9.

When upgrading from a version < 2.0 to the 2.0 version, the `bp_last_activity_migrate()` function is there to use the user meta to populate the activity table with corresponding `last_activity` activities.

We also created a repair tool in case the above upgrade task failed. This tool became obsolete but we forgot to remove it. It's now very problematic if an Admin user uses this tool as it will delete all `last_activity` activities but will fail at finding users having the `last_activity` meta.

To stop the spread, we decided to remove this tool during a minor release (12.4.0) as we felt it was an emergency.

Props vapvarun

See #9096 (trunk)



git-svn-id: https://buddypress.svn.wordpress.org/trunk@13775 cdf35c40-ae34-48e0-9cc9-0c9da1808c22