Update class-bp-activity-activity.php
When the activity item was cached and has no comments, we still need to return an empty array, just like how the rest of the function returns an empty array when there are no comments. Other functions depend on this being the case. For example, line 2618 in bp-activity-template.php in the function bp_activity_get_comment_depth will throw an error if this is not the case. We could also set bp_activity_get_comment_depth to check if $comments returned is false in that function and set it to an empty array.

But in this version, we return $comments (which is an empty array) instead of false on line 1649 (I believe) of class-bp-activity-activity.php in the public static function get_activity_comments.