When the `Extended Profiles` component is disabled, the "Profile Visibility" settings link does not 404.
The "Profile Visibility" link, `members/user/settings/profile/`, relies on the Extended Profiles component to work properly. When this component is disabled, the link should not added to the user profile settings menu.

We also did the same for the `Email` link. So that it is added only when the `Notifications` component is enabled.

Developed at https://github.com/buddypress/buddypress/pull/414/

Props joelkarunungan, nikunj8866, and vapvarun.

See #9297 (14.0)

git-svn-id: https://buddypress.svn.wordpress.org/branches/14.0@14178 cdf35c40-ae34-48e0-9cc9-0c9da1808c22